requirements.txt
fastapi
uvicorn
python-dotenv
droidrun
llama-index
google-generativeai
pydentic